+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/user/LoadMembersService.kt
@@ -40,9 +40,8 @@ class LoadMembersServiceImpl(
                     onError = { log.warn(it) { "failed loading members" } },
                 ) {
                     val room = roomStore.get(roomId).first()
-                        ?: throw IllegalStateException("room not loaded yet")
 
-                    if (!room.membersLoaded) {
+                    if (room?.membersLoaded == false) {
                         log.debug { "load members of room $roomId" }
                         val memberEvents = api.room.getMembers(
                             roomId = roomId,